Remove FormIntegrals
