+++ cpp/dolfinx/fem/petsc.cpp
@@ -62,16 +62,16 @@ la::PETScMatrix fem::create_matrix_block(
         assert(patterns[row].back());
         auto& sp = patterns[row].back();
         assert(sp);
-        const FormIntegrals<PetscScalar>& integrals = a(row, col)->integrals();
-        if (integrals.num_integrals(IntegralType::cell) > 0)
+        const fem::Form<PetscScalar>& a_ = *a(row, col);
+        if (a_.num_integrals(IntegralType::cell) > 0)
           SparsityPatternBuilder::cells(*sp, mesh->topology(), dofmaps);
-        if (integrals.num_integrals(IntegralType::interior_facet) > 0)
+        if (a_.num_integrals(IntegralType::interior_facet) > 0)
         {
           mesh->topology_mutable().create_entities(tdim - 1);
           SparsityPatternBuilder::interior_facets(*sp, mesh->topology(),
                                                   dofmaps);
         }
-        if (integrals.num_integrals(IntegralType::exterior_facet) > 0)
+        if (a_.num_integrals(IntegralType::exterior_facet) > 0)
         {
           mesh->topology_mutable().create_entities(tdim - 1);
           SparsityPatternBuilder::exterior_facets(*sp, mesh->topology(),
