+++ python/dolfinx/wrappers/fem.cpp
@@ -361,64 +361,6 @@ void fem(py::module& m)
   //           },
   //           py::return_value_policy::take_ownership);
 
-  // dolfinx::fem::FormIntegrals
-  py::class_<dolfinx::fem::FormIntegrals<PetscScalar>,
-             std::shared_ptr<dolfinx::fem::FormIntegrals<PetscScalar>>>
-      formintegrals(m, "FormIntegrals",
-                    "Holder for integral kernels and domains");
-  formintegrals
-      .def(py::init(
-          [](const std::map<dolfinx::fem::IntegralType,
-                            std::pair<std::vector<std::pair<int, py::object>>,
-                                      const dolfinx::mesh::MeshTags<int>*>>&
-                 kernels,
-             bool needs_permutation_data) {
-            using kern = std::function<void(PetscScalar*, const PetscScalar*,
-                                            const PetscScalar*, const double*,
-                                            const int*, const std::uint8_t*,
-                                            const std::uint32_t)>;
-            std::map<dolfinx::fem::IntegralType,
-                     std::pair<std::vector<std::pair<int, kern>>,
-                               const dolfinx::mesh::MeshTags<int>*>>
-                integrals;
-
-            // Loop over kernel for each entity type
-            for (auto& kernel_type : kernels)
-            {
-              // Set subdomain markers
-              integrals[kernel_type.first].second = nullptr;
-
-              // Loop over each domain kernel
-              for (auto& kernel : kernel_type.second.first)
-              {
-                auto tabulate_tensor_ptr = (void (*)(
-                    PetscScalar*, const PetscScalar*, const PetscScalar*,
-                    const double*, const int*, const std::uint8_t*,
-                    const std::uint32_t))kernel.second.cast<std::uintptr_t>();
-                integrals[kernel_type.first].first.push_back(
-                    {kernel.first, tabulate_tensor_ptr});
-              }
-            }
-            return dolfinx::fem::FormIntegrals<PetscScalar>(
-                integrals, needs_permutation_data);
-          }))
-      .def("integral_ids",
-           &dolfinx::fem::FormIntegrals<PetscScalar>::integral_ids)
-      .def_property_readonly(
-          "needs_permutation_data",
-          &dolfinx::fem::FormIntegrals<PetscScalar>::needs_permutation_data)
-      .def(
-          "integral_domains",
-          [](dolfinx::fem::FormIntegrals<PetscScalar>& self,
-             dolfinx::fem::IntegralType type, int i) {
-            const std::vector<std::int32_t>& domains
-                = self.integral_domains(type, i);
-            return py::array_t<std::int32_t>(domains.size(), domains.data(),
-                                             py::none());
-          },
-          py::return_value_policy::reference_internal,
-          "Return active domains for given integral");
-
   py::enum_<dolfinx::fem::IntegralType>(m, "IntegralType")
       .value("cell", dolfinx::fem::IntegralType::cell)
       .value("exterior_facet", dolfinx::fem::IntegralType::exterior_facet)
@@ -429,18 +371,54 @@ void fem(py::module& m)
   py::class_<dolfinx::fem::Form<PetscScalar>,
              std::shared_ptr<dolfinx::fem::Form<PetscScalar>>>(
       m, "Form", "Variational form object")
-      .def(py::init<const std::vector<std::shared_ptr<
-                        const dolfinx::function::FunctionSpace>>&,
-                    const dolfinx::fem::FormIntegrals<PetscScalar>&,
-                    const std::vector<std::shared_ptr<
-                        const dolfinx::function::Function<PetscScalar>>>&,
-                    const std::vector<std::shared_ptr<
-                        const dolfinx::function::Constant<PetscScalar>>>&,
-                    const std::shared_ptr<const dolfinx::mesh::Mesh>&>(),
+      .def(py::init([](const std::vector<std::shared_ptr<
+                           const dolfinx::function::FunctionSpace>>& spaces,
+                       const std::map<
+                           dolfinx::fem::IntegralType,
+                           std::pair<std::vector<std::pair<int, py::object>>,
+                                     const dolfinx::mesh::MeshTags<int>*>>&
+                           integrals,
+                       const std::vector<std::shared_ptr<
+                           const dolfinx::function::Function<PetscScalar>>>&
+                           coefficients,
+                       const std::vector<std::shared_ptr<
+                           const dolfinx::function::Constant<PetscScalar>>>&
+                           constants,
+                       bool needs_permutation_data,
+                       const std::shared_ptr<const dolfinx::mesh::Mesh>& mesh) {
+             using kern = std::function<void(PetscScalar*, const PetscScalar*,
+                                             const PetscScalar*, const double*,
+                                             const int*, const std::uint8_t*,
+                                             const std::uint32_t)>;
+             std::map<dolfinx::fem::IntegralType,
+                      std::pair<std::vector<std::pair<int, kern>>,
+                                const dolfinx::mesh::MeshTags<int>*>>
+                 _integrals;
+
+             // Loop over kernel for each entity type
+             for (auto& kernel_type : integrals)
+             {
+               // Set subdomain markers
+               _integrals[kernel_type.first].second = nullptr;
+
+               // Loop over each domain kernel
+               for (auto& kernel : kernel_type.second.first)
+               {
+                 auto tabulate_tensor_ptr = (void (*)(
+                     PetscScalar*, const PetscScalar*, const PetscScalar*,
+                     const double*, const int*, const std::uint8_t*,
+                     const std::uint32_t))kernel.second.cast<std::uintptr_t>();
+                 _integrals[kernel_type.first].first.push_back(
+                     {kernel.first, tabulate_tensor_ptr});
+               }
+             }
+             return dolfinx::fem::Form<PetscScalar>(
+                 spaces, _integrals, coefficients, constants,
+                 needs_permutation_data, mesh);
+           }),
            py::arg("spaces"), py::arg("integrals"), py::arg("coefficients"),
-           py::arg("constants"), py::arg("mesh") = py::none())
-      .def_property_readonly("integrals",
-                             &dolfinx::fem::Form<PetscScalar>::integrals)
+           py::arg("constants"), py::arg("need_permutation_data"),
+           py::arg("mesh") = py::none())
       .def_property_readonly("coefficients",
                              &dolfinx::fem::Form<PetscScalar>::coefficients)
       .def_property_readonly("rank", &dolfinx::fem::Form<PetscScalar>::rank)
